64bit対応

+++ nov.1/libs/cocos2d/ccMacros.h
@@ -220,7 +220,7 @@ do {															\
  On Mac it returns 1;
  On iPhone it returns 2 if RetinaDisplay is On. Otherwise it returns 1
  */
-extern float __ccContentScaleFactor;
+extern CGFloat __ccContentScaleFactor;
 #define CC_CONTENT_SCALE_FACTOR() __ccContentScaleFactor
